+++ airbyte-integrations/connectors/destination-mariadb/destination_mariadb/destination.py
@@ -62,7 +62,7 @@ class DestinationMariaDB(Destination):
         configured_catalog: ConfiguredAirbyteCatalog,
         input_messages: Iterable[AirbyteMessage],
     ) -> Iterable[AirbyteMessage]:
-        # pydevd_pycharm.settrace('192.168.178.27', port=55507, stdoutToServer=True, stderrToServer=True)
+        #pydevd_pycharm.settrace('192.168.178.27', port=55507, stdoutToServer=True, stderrToServer=True)
         logger.info("Write starting up")
         parsed_config = ConfigModel.parse_obj(config)
         self._init_sql_processor(config=parsed_config, configured_catalog=configured_catalog)
@@ -81,7 +81,7 @@ class DestinationMariaDB(Destination):
     def check(self, logger: Logger, config: Mapping[str, Any]) -> AirbyteConnectionStatus:
         _ = logger  # Unused
 
-        # pydevd_pycharm.settrace('192.168.178.27', port=55507, stdoutToServer=True, stderrToServer=True)
+        #pydevd_pycharm.settrace('192.168.178.27', port=55507, stdoutToServer=True, stderrToServer=True)
 
         try:
             parsed_config = ConfigModel.parse_obj(config)
